+++ packages/teek/index.ts
@@ -1,3 +1,4 @@
+import type { EnhanceAppContext } from "vitepress";
 import type { TeekConfig } from "@teek/config";
 import type { BaiduAnalyticsOptions, GoogleAnalyticsOptions, UmamiAnalytics } from "@teek/helper";
 import DefaultTheme from "vitepress/theme";
@@ -35,7 +36,7 @@ export * from "./version";
 export default {
   extends: DefaultTheme,
   Layout: TeekConfigProvider(TkLayout),
-  enhanceApp({ app, siteData, router }) {
+  async enhanceApp({ app, siteData, router }) {
     app.component("TkCataloguePage", TkCataloguePage);
     app.component("TkArchivesPage", TkArchivesPage);
     app.component("TkArticleOverviewPage", TkArticleOverviewPage);
@@ -51,9 +52,12 @@ export default {
     // 处理站点分析
     processSiteAnalytics(themeConfig);
     // 处理永久链接导致 404 问题
-    processPermalinkNotFoundWhenFirstLoaded({ siteData, router });
+    await processPermalinkNotFoundWhenFirstLoaded({ siteData, router });
   },
-} as DefaultThemeType & { extends: DefaultThemeType };
+} as Omit<DefaultThemeType, "enhanceApp"> & {
+  extends: DefaultThemeType;
+  enhanceApp: (options: EnhanceAppContext) => Promise<void>;
+};
 
 /**
  * 处理站点分析
@@ -78,7 +82,7 @@ const processSiteAnalytics = (themeConfig: any) => {
 /**
  * 第一次访问页面时，处理永久链接导致 404 问题
  */
-const processPermalinkNotFoundWhenFirstLoaded = ({ siteData, router }: any) => {
+const processPermalinkNotFoundWhenFirstLoaded = async ({ siteData, router }: any) => {
   const { base, cleanUrls, themeConfig } = siteData.value;
   // 404 页面处理永久链接 404 问题（仅针对首次页面刷新）
   if (router.route.path === base && router.route.data.isNotFound) {
@@ -92,11 +96,10 @@ const processPermalinkNotFoundWhenFirstLoaded = ({ siteData, router }: any) => {
     const link = cleanUrls ? decodePath : decodePath + ".html";
     const filePath = themeConfig.permalinks.inv[link];
 
-    // 如果通过永久链接获取的文件路径存在，则跳转
+    // 通过永久链接获取的文件路径存在，则跳转
     if (filePath) {
       const targetUrl = base + filePath + search + hash;
-      router.go(targetUrl);
-      return false;
+      await router.go(targetUrl);
     }
   }
 };